[text]: allow materials to have both color and texture
Previously the diffuse visual for a material could
only have a texture or a color. Allow visuals to support
both, they simply get multiplied together. This will
enable us to support glyph textures getting multiplied
by text color.

Some other minor cleanup:
+ fixed minor bug in video texture where seconds was an int
+ added some virtual destructors
+ use the name of materials when debugging

+++ ios/ViroKit/VROVideoTextureiOS.h
@@ -52,7 +52,7 @@ public:
     void pause();
     void play();
     bool isPaused();
-    void seekToTime(int seconds);
+    void seekToTime(float seconds);
     
     void setMuted(bool muted);
     void setVolume(float volume);
